add parent class component.base for others components
