+++ src/component.base.js
@@ -1,0 +1,31 @@
+// component.base.js
+
+import {createElement} from './utils';
+
+
+export default class Component {
+  constructor() {
+    if (new.target === Component) {
+      throw new Error(`Can't instantiate Component, only concrete one.`);
+    }
+  }
+
+  get template() {
+    throw new Error(`You have to define template`);
+  }
+
+  bind() {}
+
+  unbind() {}
+
+  render() {
+    this._element = createElement(this.template);
+    this.bind();
+    return this._element;
+  }
+
+  unrender() {
+    this.unbind();
+    this._element = null;
+  }
+}
