--- src/create-element.js
@@ -1,8 +1,0 @@
-// create-element.js
-
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
